+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 Key: [+] added, [-] removed, [~] modified, [!] important.
 
 ## 202010r1
-+ [!] Remove/Patch code the uses privileged intents.
++ [!] Remove/Patch code the uses privileged intents. +1
++ [!] Removed references to guild members in database.
 + [~] Updated dependencies.
 
 ## 202009r1